Fix compilation problems with newer toolchains.
The biggest change is removing the use of readdir_r which is deprecated.
We didn't really need it, because each thred opens the directory itself.
According to the readdir_r man page on Linux, "However, in modern
implementations (including the glibc implementation), concurrent calls to
readdir(3) that specify different directory streams  are  thread-safe."

--HG--
branch : de-bitrot-2018

+++ moss_serv.cc
@@ -530,7 +530,7 @@ void * serv_main(void *serv) {
   Server::Connection *conn;
 
   // local state
-  int exit_value = 1;
+  intptr_t exit_value = 1;
   int i, j;
   struct iovec *iov = NULL;
   u_int wrote;
